==> notebooks/2_Preprocess_data.ipynb <==
# В этом ноутбуке будут рассмотрены различные видны препроцессинга данных для моделей классификации

import pandas as pd
import numpy as np
from tqdm import tqdm
from sklearn.neighbors import KDTree
import sys

sys.path.append('/Users/alexander.savelyev/PycharmProjects/Snow_recognition')
from preprocess import Preprocess

preprocess.py -- в этом файле написаны все методы препроцессинга, которые будут описываться в этой тетрадке 

preprocessor = Preprocess()

df = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data.csv")

df.head(5)

### 1. Простой препроцессинг

Для каждой сцены будем в ней нормализовывать с помощью min_max в ней координаты x, y, z, intensity и ring

df_1 = preprocessor.simple_preprocess_normalize_all_columns(df)
df_1.to_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns.csv")
df_1 = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns.csv", index_col=0)

df_1

Проверим что координаты действительно нормализованы

df_1.describe()

Как видим, x_normalized, y_normalized, z_normalized, ring_normailized и intensity_normailized находяться от 0 до 1

Будем использовать эти данные для обучения логистической регресии, потому что линейная модель может быть очень подвержена большой высокоими значениями фичей

### 2. Препроцессинг на основе KDTree

Для каждой сцены будем делать следующее:

1) Нормализовывать координаты x, y и z, а затем сохранять их в x_normalized, y_normalized и z_normalized.

2) Для всех точек сцены по кординатам x_normalized, y_normalized и z_normalized будем строить KDTree. KDTree -- структура для быстрого поиска ближайших соседей, будем использовать ее потому что полный перебор очень долго делается.

3) Затем для каждой точки сцены будем находить ее 7 ближайших соседей с помощью KDTree. 

4) Будем считать следующие характеристики на основе ее ближайших соседей:

    * dist_mean -- среднее расстояние от точки до всех ее соседей
    * dist_std -- стандартное отклонение расстояний от точки до всех ее соседей
    * dist_for_first -- расстоние до самой ближней точки (из найденных соседей)
    * dist_for_last -- расстоние до самой дальней точки точки (из найденных соседей)
    
    * ring_mean -- средний номер луча от точки до всех ее соседей
    * ring_std -- стандартное отклонение номера луча от точки до всех ее соседей
    * ring_for_first -- номера луча самой ближней точки (из найденных соседей)
    * ring_for_last -- номера луча самой дальней точки точки (из найденных соседей)
    
    * int_mean -- среднея интенсивность от точки до всех ее соседей
    * int_std -- стандартное отклонение интенсивности от точки до всех ее соседей
    * int_for_first -- интенсивность самой ближней точки (из найденных соседей)
    * int_for_last -- интенсивность самой дальней точки точки (из найденных соседей)
    

Такой фичи инжеринг обуслевден тем фактом, что проведя EDA мы сделали несколько выводов:

* Также дисперсия для снега меньше, чем у обычных точек, из этого можно сделать вывод, что снег скапливается в одном месте, а не "раскидан" по всей сцене.

* По этой информации можно сделать вывод, что снег в среднем находиться выше, чем другие точки, потому z_normailized больше для label = 1, чем z_normailized для label = 0.

* По этой информации можно сделать вывод, что снег в среднем намного тусклее, и эта тусклость находиться около 0.

Значит если точка находиться в окружении точек снега, то скорее всего она будет тоже снегом, эту гипотезу мы проверем в будущем

df_kdtree = preprocessor.kdtree_preprocess(df)
df_kdtree.to_csv("/Volumes/HP P800/itmo/Lidar data/data_kdtree_preprocess.csv")
df_kdtree = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_kdtree_preprocess.csv")

df_kdtree
