==> notebooks/3 Train Test split.ipynb <==
# В этой тетрадке мы разобъем все три датасета на train и test

import pandas as pd
import numpy as np

230 -- количество сцен в наших данных

scenes = np.linspace(0, 229, num=230).astype(int)
np.random.shuffle(scenes)

scenes

возьмем 80% всех сцен в train и 20% в test

train_size = int(0.8 * 230)

train_scene_id = scenes[0 : train_size]
test_scene_id = scenes[train_size : ]

train_scene_id

test_scene_id

df = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data.csv")

df_train = df[df['scene_id'].isin(train_scene_id)]
df_test = df[df['scene_id'].isin(test_scene_id)]

df_train.to_csv("/Volumes/HP P800/itmo/Lidar data/data_train.csv")
df_test.to_csv("/Volumes/HP P800/itmo/Lidar data/data_test.csv")

df_1 = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns.csv", index_col=0)

df_1_train = df_1[df_1['scene_id'].isin(train_scene_id)]
df_1_test = df_1[df_1['scene_id'].isin(test_scene_id)]

df_1_train.to_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns_train.csv")
df_1_test.to_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns_test.csv")

df_kdtree = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_kdtree_preprocess.csv", index_col=0)

df_kdtree_train = df_kdtree[df_kdtree['scene_id'].isin(train_scene_id)]
df_kdtree_test = df_kdtree[df_kdtree['scene_id'].isin(test_scene_id)]

df_kdtree_train.to_csv("/Volumes/HP P800/itmo/Lidar data/df_kdtree_train.csv")
df_kdtree_test.to_csv("/Volumes/HP P800/itmo/Lidar data/df_kdtree_test.csv")
