==> notebooks/4 logistic regression.ipynb <==
# В этом ноутбуке будем тренировать логистическую регререссию на наших данных

import pandas as pd
import numpy as np
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score

будем использовать нормализованный датасет

#df_train = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_train.csv")
df_train =  pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns_train.csv")
# df_train =  pd.read_csv("/Volumes/HP P800/itmo/Lidar data/df_kdtree_train.csv")

Сделаем проверку на дисбаланс классов

df_train.groupby(["label"]).count()

Как видим дисбаланас классов очень большой, избавимся от него путем раскопирования данных

k = df_train[df_train["label"] == 1.0]

m = len(df_train[df_train["label"] == 0.0]) // len(df_train[df_train["label"] == 1.0])
ost = len(df_train[df_train["label"] == 0.0]) % len(df_train[df_train["label"] == 1.0])

for i in range(0,m-1):
    df_train = df_train.append(k)

df_train = df_train.append(k.head(ost))

df_train.groupby(["label"]).count()

Теперь количество сэмлов в классах равны

Разделиим данные на X и y. Не будем использовать фичи z и ring вместе, потому что они имеют большую корреляцию

X_train = df_train[["x_normailized", "y_normailized", "z_normailized", "intensity_normailized"]]
y_train = df_train[["label"]]

clf = LogisticRegression(random_state=0).fit(X_train, y_train)

Выведем roc_auc_score для тренировочных данных

roc_auc_score(y_train, clf.predict_proba(X_train)[:, 1])

Теперь оценим результат roc_auc_score на тестовых данных

#df_test = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_test.csv")
df_test = pd.read_csv("/Volumes/HP P800/itmo/Lidar data/data_simple_preprocess_normalize_all_columns_test.csv")
# df_test = pd.read_csv(("/Volumes/HP P800/itmo/Lidar data/df_kdtree_test.csv")

X_test = df_test[["x_normailized", "y_normailized", "z_normailized", "intensity_normailized"]] 
y_test = df_test[["label"]]

roc_auc_score(y_test, clf.predict_proba(X_test)[:, 1])

Посмотрим на веса классификатора

print(clf.coef_)

Как видим z_normailized имеет наибольший вес для данной модели, тогда если значение z большое, то классификатор отнесете его к снежинке с большой вероятностью
